+++ cmd/mnemos/consolidate.go
@@ -15,7 +15,8 @@ import (
 //
 //	mnemos consolidate [--dry-run] [--forget-below-trust <f>] [--forget-refuted]
 //	                   [--reinforce-validated] [--reinforce-playbooks] [--credit]
-//	                   [--salience] [--plastic] [--synthesize] [--replay-top-k <n>]
+//	                   [--salience] [--plastic] [--decay-associations]
+//	                   [--synthesize] [--replay-top-k <n>]
 //
 // --plastic engages the ADR-0015 adaptive learning-rate controls on the credit stage
 // (metaplasticity + neuromodulation); MNEMOS_PLASTICITY_SENSITIVITY tunes the latter.
@@ -84,6 +85,8 @@ func parseConsolidateOpts(args []string, f Flags) (mnemos.ConsolidateOptions, er
 			opts.AssignSalience = true
 		case "--plastic":
 			opts.Plastic = true
+		case "--decay-associations":
+			opts.DecayAssociations = true
 		case "--reinforce-playbooks":
 			opts.ReinforcePlaybooks = true
 		case "--synthesize":